Use char instead of Char, fix GC bug
The call to NEW_STRING could trigger a GC and thus invalidate ptr.
Fix this by reordering commands. Also use char instead of Char, so
that the GAP kernel can eventually get rid of the latter. Finally,
switch ptr to a const pointer, as we are only reading from it.